dataserverapp/requirement.txt
fastapi
uvicorn
pandas
langchain
langchain-community
langchain-core
langchainhub
langchain-google-genai
python-dotenv
dnspython 
langgraph
python-multipart
certifi
supabase
dnspython 
google.generativeai
pandasql
openpyxl
